brillig_vm
Module foreign_call

Implementation for foreign calls

Functions

vector_element_size 🔒

Returns the total number of field elements required to represent the elements in the vector in memory.