brillig_vm
Struct Memory

pub struct Memory<F> {
    inner: Vec<MemoryValue<F>>,
}

The VMโs memory. Memory is internally represented as a vector of values. We grow the memory when values past the end are set, extending with 0s.

Fields
inner: Vec<MemoryValue<F>>
Implementations

impl<F: AcirField> Memory<F>

fn get_stack_pointer(&self) -> u32

Read the value from slot 0.

Panics if itโs not a U32.

fn resolve(&self, address: MemoryAddress) -> u32

Resolve an address to either:

itself, if itโs a direct address, or

the current stack pointer plus the offset, if itโs relative.

Returns a memory slot index.

pub fn read(&self, address: MemoryAddress) -> MemoryValue<F>

Reads the numeric value at the address.

If the address is beyond the size of memory, a default value is returned.

pub fn read_ref(&self, ptr: MemoryAddress) -> MemoryAddress

Reads the value at the address and returns it as a direct memory address, without dereferencing the pointer itself to a numeric value.

pub fn write_ref(&mut self, ptr: MemoryAddress, address: MemoryAddress)

Sets ptr to point at address.

pub fn read_slice( &self, address: MemoryAddress, len: usize, ) -> &[MemoryValue<F>]

Read a contiguous vector of memory starting at address, up to len slots.

Panics if the end index is beyond the size of the memory.

pub fn write(&mut self, address: MemoryAddress, value: MemoryValue<F>)

Sets the value at address to value

fn resize_to_fit(&mut self, size: usize)

Increase the size of memory fit size elements, or the current length, whichever is bigger.

pub fn write_slice(&mut self, address: MemoryAddress, values: &[MemoryValue<F>])

Sets the values after address to values

pub fn values(&self) -> &[MemoryValue<F>]

Returns the values of the memory