Crate brillig_vm

The Brillig VM is a specialized VM which allows the ACVM to perform custom non-determinism.

Brillig bytecode is distinct from regular ACIR in that it does not generate constraints.

Re-exports

pub use crate::fuzzing::BranchToFeatureMap;

pub use acir::brillig;

Modules

arithmetic 🔒

Implementations for [binary field operations][acir::brillig::Opcode::BinaryFieldOp] and [binary integer operations][acir::brillig::Opcode::BinaryIntOp].

black_box 🔒

Implementations for VM native [black box functions][acir::brillig::Opcode::BlackBox].

cast 🔒

Implementation for the [cast operation][acir::brillig::Opcode::Cast].

foreign_call 🔒

Implementation for [foreign calls][acir::brillig::Opcode::ForeignCall]

fuzzing

memory 🔒

Implementation of the VM’s memory.

offsets

Offset constants for arrays and vectors:

Structs

BrilligProfilingSample

A sample for an executed opcode.

Memory

The VM’s memory. Memory is internally represented as a vector of values. We grow the memory when values past the end are set, extending with 0s.

VM

VM encapsulates the state of the Brillig VM during execution.

Enums

FailureReason

Represents the reason why the Brillig VM failed during execution.

MemoryValue

A single typed value in the Brillig VM’s memory.

VMStatus

Represents the current execution status of the Brillig VM.

Constants

FREE_MEMORY_POINTER_ADDRESS

The free memory pointer is always in slot 1.

MEMORY_ADDRESSING_BIT_SIZE

The bit size used for addressing memory within the Brillig VM.

STACK_POINTER_ADDRESS

The current stack pointer is always in slot 0.

Type Aliases

BrilligProfilingSamples

All samples for each opcode that was executed.

ErrorCallStack

The error call stack contains the opcode indexes of the call stack at the time of failure, plus the index of the opcode that failed.

NextOpcodePositionOrState

The position of the next opcode that will be executed in the bytecode, or an id of a specific state produced by the opcode.

OpcodePosition

The position of an opcode that is currently being executed in the bytecode.