brillig_vm
Module fuzzing

Structs

FuzzingTrace 🔒

Context structure for all information necessary to compute the fuzzing trace

Constants

FUZZING_COMPARISON_FALSE_STATE 🔒

A state that represents a false comparison as part of a feature

FUZZING_COMPARISON_LOG_RANGE_START_STATE 🔒

The start of the range of the states that represent logarithm of the difference between the comparison arguments as part of a feature

FUZZING_COMPARISON_TRUE_STATE 🔒

A state that represents a true comparison as part of a feature

Type Aliases

Branch

A tuple of the current opcode position and the next opcode position or state

BranchToFeatureMap

A map for translating encountered branching logic to features for fuzzing

UniqueFeatureIndex

The index of a unique feature in the fuzzing trace